pyproject.toml
[tool.poetry]
name = "picmeo"
version = "0.1.0"
description = ""
authors = ["gonzalogacc <gonzalogacc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.24.1"
pytest = "^7.4.1"
python-dotenv = "^1.0.0"
google-cloud-storage = "^2.12.0"
sqlmodel = "^0.0.14"
pydantic = "^2.5.3"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
